feat: implement comprehensive ERP module and multilingual localization support with core backend enhancements

+++ database/migrations/2023_01_15_335340_create_user_referral_commissions_table.php
@@ -13,28 +13,30 @@ class CreateUserReferralCommissionsTable extends Migration
      */
     public function up()
     {
-        Schema::create('user_referral_commissions', function (Blueprint $table) {
-            $table->id();
-
-            $table->bigInteger('user_id')->unsigned()->nullable();
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-
-            $table->bigInteger('referred_user_id')->unsigned()->nullable();
-            $table->foreign('referred_user_id')->references('id')->on('users')->onDelete('cascade');
-
-            $table->bigInteger('referred_invoice_id')->unsigned()->nullable();
-            $table->foreign('referred_invoice_id')->references('id')->on('invoices')->onDelete('cascade');
-
-            $table->bigInteger('currency')->unsigned()->index();
-            $table->foreign('currency')->references('id')->on('currencies')->onDelete('cascade');
-
-            $table->double('amount', 23, 3)->default('0');
-
-            $table->date('convert_to_balance_on')->nullable();
-
-            $table->timestamps();
-            $table->softDeletes();
-        });
+        if (!Schema::hasTable('user_referral_commissions')) {
+            Schema::create('user_referral_commissions', function (Blueprint $table) {
+                $table->id();
+    
+                $table->bigInteger('user_id')->unsigned()->nullable();
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+    
+                $table->bigInteger('referred_user_id')->unsigned()->nullable();
+                $table->foreign('referred_user_id')->references('id')->on('users')->onDelete('cascade');
+    
+                $table->bigInteger('referred_invoice_id')->unsigned()->nullable();
+                $table->foreign('referred_invoice_id')->references('id')->on('invoices')->onDelete('cascade');
+    
+                $table->bigInteger('currency')->unsigned()->index();
+                $table->foreign('currency')->references('id')->on('currencies')->onDelete('cascade');
+    
+                $table->double('amount', 23, 3)->default('0');
+    
+                $table->date('convert_to_balance_on')->nullable();
+    
+                $table->timestamps();
+                $table->softDeletes();
+            });
+        }
     }
 
     /**
